backend/requirements.txt
fastapi==0.103.1
pymongo==4.5.0
uvicorn==0.23.2
python-decouple==3.8
pyjwt==2.8.0
bcrypt==4.0.1
pydantic[email]==2.4.2
passlib==1.7.4
debugpy==1.8.0
phonenumbers==8.13.25
motor==3.3.2
pytest==7.4.3
requests==2.31.0
python-decouple==3.8
httpcore==1.0.2
httpx==0.25.1
pytest-asyncio==0.21.1
pytest-order==1.2.0
aiohttp==3.9.0